# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Rihib API
  description: Personal website API for articles and profile data
  version: 1.0.0
  contact:
    name: Rihito Bannai
    url: https://rihib.dev
servers:
  - url: http://localhost:8787
    description: Development server
  - url: https://rihib.dev
    description: Production server

paths:
  /:
    get:
      summary: Health check endpoint
      description: Returns basic API information
      responses:
        '200':
          description: API is running
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Rihib API Server"

  /api:
    get:
      summary: API root endpoint
      description: Returns API information
      responses:
        '200':
          description: API information
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Rihib API Server"

  /api/articles:
    get:
      summary: Get articles by type and locale
      description: Retrieve blog posts or news items filtered by locale
      parameters:
        - name: locale
          in: query
          required: true
          schema:
            $ref: '#/components/schemas/Locale'
          description: Language locale
        - name: type
          in: query
          required: true
          schema:
            $ref: '#/components/schemas/ArticleType'
          description: Article type
      responses:
        '200':
          description: Articles retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArticlesResponse'
        '400':
          description: Invalid parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/blog:
    get:
      summary: Get blog posts
      description: Retrieve blog posts filtered by locale
      parameters:
        - name: locale
          in: query
          required: true
          schema:
            $ref: '#/components/schemas/Locale'
          description: Language locale
      responses:
        '200':
          description: Blog posts retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArticlesResponse'
        '400':
          description: Invalid parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/news:
    get:
      summary: Get news items
      description: Retrieve news items filtered by locale
      parameters:
        - name: locale
          in: query
          required: true
          schema:
            $ref: '#/components/schemas/Locale'
          description: Language locale
      responses:
        '200':
          description: News items retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArticlesResponse'
        '400':
          description: Invalid parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

components:
  schemas:
    Locale:
      type: string
      enum: [en, ja]
      description: Language locale
      example: en

    ArticleType:
      type: string
      enum: [blog, news]
      description: Type of article
      example: blog

    Article:
      type: object
      required:
        - id
        - title
        - published_at
        - url
        - type
        - locale
        - created_at
      properties:
        id:
          type: integer
          format: int32
          minimum: 1
          description: Unique article identifier
          example: 1
        title:
          type: string
          minLength: 1
          maxLength: 500
          description: Article title
          example: "My First Blog Post"
        published_at:
          type: string
          format: date
          pattern: '^\d{4}-\d{2}-\d{2}$'
          description: Publication date in YYYY-MM-DD format
          example: "2024-01-15"
        url:
          type: string
          format: uri
          maxLength: 2000
          description: Article URL
          example: "https://example.com/my-first-post"
        type:
          $ref: '#/components/schemas/ArticleType'
        locale:
          $ref: '#/components/schemas/Locale'
        created_at:
          type: string
          format: date-time
          description: Creation timestamp in ISO format with timezone
          example: "2024-01-15T10:30:00Z"

    ArticlesResponse:
      type: object
      required:
        - articles
      properties:
        articles:
          type: array
          items:
            $ref: '#/components/schemas/Article'
          maxItems: 1000
          description: Array of articles
          example:
            - id: 1
              title: "My First Blog Post"
              published_at: "2024-01-15"
              url: "https://example.com/my-first-post"
              type: "blog"
              locale: "en"
              created_at: "2024-01-15T10:30:00Z"

    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          minLength: 1
          maxLength: 1000
          description: Error message
          example: "Internal server error"

    CreateArticle:
      type: object
      required:
        - title
        - published_at
        - url
        - type
        - locale
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 500
          description: Article title
          example: "New Article Title"
        published_at:
          type: string
          format: date
          pattern: '^\d{4}-\d{2}-\d{2}$'
          description: Publication date in YYYY-MM-DD format
          example: "2024-01-15"
        url:
          type: string
          format: uri
          maxLength: 2000
          description: Article URL
          example: "https://example.com/new-article"
        type:
          $ref: '#/components/schemas/ArticleType'
        locale:
          $ref: '#/components/schemas/Locale'

    UpdateArticle:
      type: object
      required:
        - id
      properties:
        id:
          type: integer
          format: int32
          minimum: 1
          description: Article ID to update
          example: 1
        title:
          type: string
          minLength: 1
          maxLength: 500
          description: Article title
          example: "Updated Article Title"
        published_at:
          type: string
          format: date
          pattern: '^\d{4}-\d{2}-\d{2}$'
          description: Publication date in YYYY-MM-DD format
          example: "2024-01-15"
        url:
          type: string
          format: uri
          maxLength: 2000
          description: Article URL
          example: "https://example.com/updated-article"
        type:
          $ref: '#/components/schemas/ArticleType'
        locale:
          $ref: '#/components/schemas/Locale'